requirements.txt
nose==1.3.1
rednose
salt==2014.1
Fabric==1.8.1
apache-libcloud >=0.14.0-beta3
lxml
pyyaml
GitPython==0.3.2.RC1
git+http://github.com/ministryofjustice/cotton.git#egg=cotton>=0.2.0
